pyproject.toml
[tool.poetry]
name = "vasthoodclean"
version = "1.0"
description = ""
authors = ["posholokia <ilya.posholokk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-ninja = "^1.2.1"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
psycopg2 = "^2.9.9"
django-nested-admin = "^4.0.2"
django-admin-interface = "^0.28.8"
punq = "^0.7.0"
django-cors-headers = "^4.4.0"
twilio = "^9.2.3"
pydantic-settings = "^2.4.0"
loguru = "^0.7.2"
redis = "^5.0.8"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.1"
isort = "^5.13.2"
pre-commit = "^3.7.1"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".ci",
    ".ruff_cache",
    ".venv",
    "venv",
    "test",
    ".mypy_cache",
    "src/config/settings/__init__.py",
    ]

line-length = 79
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
ignore = ["E712"]

[tool.isort]
order_by_type = true
force_sort_within_sections=false
multi_line_output=3
include_trailing_comma=true
line_length=79
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
length_sort_straight=true
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["core", ]
skip_glob=[
    ".git",
    ".idea",
    ".ci",
    ".ruff_cache",
    "venv",
    ".mypy_cache"
]
default_section = "THIRDPARTY"
known_firstparty=["core"]

sections=[
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]